Wilson: Moved 'IServerEvent' and 'IClientEvent' generics in 'common'
