+++ src/wilson/mappers/jsonStringToClientEvent.ts
@@ -1,5 +1,5 @@
 import InvalidClientEventException from "../exceptions/InvalidClientEventException";
-import ClientEvent from "../types/ClientEvent";
+import ClientEvent from "../types/common/events/ClientEvent";
 
 export default function jsonStringToClientEvent(json: string): ClientEvent {
