+++ src/wilson/types/common/events/ClientEvent.ts
@@ -1,8 +1,11 @@
 import EClientAction from "../../../enums/EClientAction";
+import IClientEvent from "./IClientEvent";
+import ClientNewConversationInviteAckEventPayload from "./payloads/client/ClientNewConversationInviteAckEventPayload";
+import ClientNewConversationInviteEventPayload from "./payloads/client/ClientNewConversationInviteEventPayload";
 
-interface IClientEvent<T, U extends EClientAction> {
-    action: U,
-    payload: T
-}
+export type ClientNewConversationInviteEvent = IClientEvent<ClientNewConversationInviteEventPayload, EClientAction.NEW_CONVERSATION_INVITE>;
+export type ClientNewConversationInviteAckEvent = IClientEvent<ClientNewConversationInviteAckEventPayload, EClientAction.NEW_CONVERSATION_INVITE_ACK>;
 
-export default IClientEvent;
+type ClientEvent = ClientNewConversationInviteEvent | ClientNewConversationInviteAckEvent;
+
+export default ClientEvent;
