+++ src/wilson/types/common/events/IClientEvent.ts
@@ -1,0 +1,8 @@
+import EClientAction from "../../../enums/EClientAction";
+
+interface IClientEvent<T, U extends EClientAction> {
+    action: U,
+    payload: T
+}
+
+export default IClientEvent;
