+++ src/wilson/types/common/events/IServerEvent.ts
@@ -1,0 +1,9 @@
+import EServerAction from "../../../enums/EServerAction";
+
+interface IServerEvent<T, U extends EServerAction> {
+    to: string,
+    action: U,
+    payload: T
+}
+
+export default IServerEvent;
